# attendance-backend-main/public/swagger.json
{
	"openapi": "3.0.0",
	"components": {
...
		"schemas": {
...
			"ServiceResponse_null_": {
				"properties": {
					"success": {
						"type": "boolean"
					},
					"message": {
						"type": "string"
					},
					"result": {
						"anyOf": [],
						"nullable": true
					},
					"statusCode": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"success",
					"message",
					"result",
					"statusCode"
				],
				"type": "object",
				"additionalProperties": true
			},
...
			"UserStatus": {
				"type": "string",
				"enum": [
					"active",
					"inactive",
					"suspended",
					"pending",
					"deleted",
					"archived",
					"rejected",
					"blacklisted"
				]
			},
...
			"FaceVerifyRequest": {
				"properties": {
					"department": {
						"type": "string"
					},
					"faceImageBase64": {
						"type": "string"
					},
					"apiKey": {
						"type": "string"
					}
				},
				"type": "object"
			},
			"FaceVerifyResult": {
				"properties": {
					"profilePhoto": {
						"type": "string"
					},
					"department": {
						"type": "string"
					},
					"fullName": {
						"type": "string"
					},
					"userId": {
						"type": "string"
					},
					"entityType": {
						"type": "string",
						"enum": [
							"employee",
							"visitor",
							"unknown"
						]
					},
					"matched": {
						"type": "boolean"
					}
				},
				"required": [
					"matched"
				],
				"type": "object"
			},
			"FaceVerifyError": {
				"properties": {
					"error": {
						"type": "string"
					},
					"matched": {
						"type": "boolean",
						"enum": [
							false
						],
						"nullable": false
					}
				},
				"required": [
					"matched"
				],
				"type": "object"
			},
			"ServiceResponse_FaceVerifyResult-or-FaceVerifyError_": {
				"properties": {
					"success": {
						"type": "boolean"
					},
					"message": {
						"type": "string"
					},
					"result": {
						"anyOf": [
							{
								"$ref": "#/components/schemas/FaceVerifyResult"
							},
							{
								"$ref": "#/components/schemas/FaceVerifyError"
							}
						],
						"nullable": true
					},
					"statusCode": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"success",
					"message",
					"result",
					"statusCode"
				],
				"type": "object",
				"additionalProperties": true
			},
...
			"LoginRequest": {
				"properties": {
					"email": {
						"type": "string"
					},
					"password": {
						"type": "string"
					},
					"phone": {
						"type": "string"
					},
					"type": {
						"type": "string",
						"enum": [
							"phone",
							"email"
						]
					}
				},
				"required": [
					"password"
				],
				"type": "object",
				"additionalProperties": true
			},
...
			"IUserResponse": {
				"properties": {
					"id": {
						"type": "string"
					},
					"fullName": {
						"type": "string"
					},
					"email": {
						"type": "string"
					},
					"phoneNumber": {
						"anyOf": [
							{
								"type": "string"
							},
							{
								"type": "number",
								"format": "double"
							}
						]
					},
					"status": {
						"$ref": "#/components/schemas/UserStatus"
					},
					"roles": {
						"items": {},
						"type": "array"
					},
					"profilePicture": {
						"type": "string"
					},
					"dateOfBirth": {
						"type": "string",
						"format": "date-time"
					},
					"nationalId": {
						"type": "string"
					},
					"fingerPrint": {
						"type": "string"
					},
					"face": {
						"type": "string"
					},
					"voice": {
						"type": "string"
					},
					"signature": {
						"type": "string"
					},
					"passport": {
						"type": "string"
					},
					"motion": {
						"type": "string"
					},
					"igipande": {
						"type": "string"
					},
					"ocr": {
						"type": "string"
					},
					"gesture": {
						"type": "string"
					},
					"pupil": {
						"type": "string"
					},
					"otherBiometric": {
						"type": "string"
					},
					"nationality": {
						"type": "string"
					},
					"country": {
						"type": "string"
					},
					"province": {
						"type": "string"
					},
					"district": {
						"type": "string"
					},
					"sector": {
						"type": "string"
					},
					"cell": {
						"type": "string"
					},
					"village": {
						"type": "string"
					},
					"personalInChargeNationalId": {
						"type": "string"
					},
					"personalInChargeName": {
						"type": "string"
					},
					"personalInChargePhone": {
						"anyOf": [
							{
								"type": "string"
							},
							{
								"type": "number",
								"format": "double"
							}
						]
					},
					"personalInChargeEmail": {
						"type": "string"
					},
					"personalInChargeRelation": {
						"type": "string"
					},
					"personalInChargeOtherDetails": {
						"type": "string"
					},
					"createdAt": {
						"type": "string",
						"format": "date-time"
					},
					"updatedAt": {
						"type": "string",
						"format": "date-time"
					}
				},
				"required": [
					"id",
					"fullName",
					"email",
					"status"
				],
				"type": "object",
				"additionalProperties": true
			},
			"ServiceResponse__user-IUserResponse-or-null__": {
				"properties": {
					"success": {
						"type": "boolean"
					},
					"message": {
						"type": "string"
					},
					"result": {
						"properties": {
							"user": {
								"allOf": [
									{
										"$ref": "#/components/schemas/IUserResponse"
									}
								],
								"nullable": true
							}
						},
						"required": [
							"user"
						],
						"type": "object",
						"nullable": true
					},
					"statusCode": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"success",
					"message",
					"result",
					"statusCode"
				],
				"type": "object",
				"additionalProperties": true
			},
			"SignupRequest": {
				"properties": {
					"fullName": {
						"type": "string"
					},
					"email": {
						"type": "string"
					},
					"roleType": {
						"type": "string"
					},
					"password": {
						"type": "string"
					},
					"phone": {
						"type": "string"
					}
				},
				"required": [
					"fullName",
					"password",
					"phone"
				],
				"type": "object",
				"additionalProperties": true
			},
			"ServiceResponse__user-IUserResponse-or-null--accessToken_63_-string--refreshToken_63_-string__": {
				"properties": {
					"success": {
						"type": "boolean"
					},
					"message": {
						"type": "string"
					},
					"result": {
						"properties": {
							"refreshToken": {
								"type": "string"
							},
							"accessToken": {
								"type": "string"
							},
							"user": {
								"allOf": [
									{
										"$ref": "#/components/schemas/IUserResponse"
									}
								],
								"nullable": true
							}
						},
						"required": [
							"user"
						],
						"type": "object",
						"nullable": true
					},
					"statusCode": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"success",
					"message",
					"result",
					"statusCode"
				],
				"type": "object",
				"additionalProperties": true
			},
			"ServiceResponse__exists-boolean__": {
				"properties": {
					"success": {
						"type": "boolean"
					},
					"message": {
						"type": "string"
					},
					"result": {
						"properties": {
							"exists": {
								"type": "boolean"
							}
						},
						"required": [
							"exists"
						],
						"type": "object",
						"nullable": true
					},
					"statusCode": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"success",
					"message",
					"result",
					"statusCode"
				],
				"type": "object",
				"additionalProperties": true
			},
			"ServiceResponse__accessToken-string--refreshToken-string__": {
				"properties": {
					"success": {
						"type": "boolean"
					},
					"message": {
						"type": "string"
					},
					"result": {
						"properties": {
							"refreshToken": {
								"type": "string"
							},
							"accessToken": {
								"type": "string"
							}
						},
						"required": [
							"refreshToken",
							"accessToken"
						],
						"type": "object",
						"nullable": true
					},
					"statusCode": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"success",
					"message",
					"result",
					"statusCode"
				],
				"type": "object",
				"additionalProperties": true
			},
...
		},
		"securitySchemes": {
			"jwt": {
				"type": "apiKey",
				"name": "Authorization",
				"in": "header",
				"description": "JWT token (Bearer <token>)"
			}
		}
	},
	"info": {
		"title": "Attendance Backend API",
		"version": "1.0.0",
		"description": "API documentation for attendance management system",
		"license": {
			"name": "ISC"
		},
		"contact": {
			"name": "API Support"
		}
	},
	"paths": {
...
		"/api/faces/verify": {
			"post": {
				"operationId": "VerifyFace",
				"responses": {
					"200": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ServiceResponse_FaceVerifyResult-or-FaceVerifyError_"
								}
							}
						}
					},
					"204": {
						"description": "No content"
					},
					"400": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ServiceResponse_FaceVerifyResult-or-FaceVerifyError_"
								}
							}
						}
					}
				},
				"tags": [
					"Faces"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/FaceVerifyRequest"
							}
						}
					}
				}
			}
		},
...
		"/api/auth/login": {
			"post": {
				"operationId": "Login",
				"responses": {
					"200": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ServiceResponse__user-IUserResponse-or-null__"
								}
							}
						}
					},
					"401": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ServiceResponse__user-IUserResponse-or-null__"
								}
							}
						}
					},
					"403": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ServiceResponse__user-IUserResponse-or-null__"
								}
							}
						}
					}
				},
				"tags": [
					"Authentication"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/LoginRequest"
							}
						}
					}
				}
			}
		},
		"/api/auth/me": {
			"get": {
				"operationId": "GetCurrentUser",
				"responses": {
					"200": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ServiceResponse__user-IUserResponse-or-null__"
								}
							}
						}
					},
					"401": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ServiceResponse__user-IUserResponse-or-null__"
								}
							}
						}
					}
				},
				"tags": [
					"Authentication"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": []
			}
		},
		"/api/auth/signup": {
			"post": {
				"operationId": "Signup",
				"responses": {
					"201": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ServiceResponse__user-IUserResponse-or-null--accessToken_63_-string--refreshToken_63_-string__"
								}
							}
						}
					},
					"400": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ServiceResponse__user-IUserResponse-or-null--accessToken_63_-string--refreshToken_63_-string__"
								}
							}
						}
					},
					"409": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ServiceResponse__user-IUserResponse-or-null--accessToken_63_-string--refreshToken_63_-string__"
								}
							}
						}
					}
				},
				"tags": [
					"Authentication"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/SignupRequest"
							}
						}
					}
				}
			}
		},
		"/api/auth/check-email": {
			"post": {
				"operationId": "CheckEmail",
				"responses": {
					"200": {
						"description": "Email check Successfull",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ServiceResponse__exists-boolean__"
								}
							}
						}
					}
				},
				"tags": [
					"Authentication"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"properties": {
									"email": {
										"type": "string"
									}
								},
								"required": [
									"email"
								],
								"type": "object"
							}
						}
					}
				}
			}
		},
		"/api/auth/refresh": {
			"post": {
				"operationId": "RefreshToken",
				"responses": {
					"200": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ServiceResponse__accessToken-string--refreshToken-string__"
								}
							}
						}
					},
					"401": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ServiceResponse__accessToken-string--refreshToken-string__"
								}
							}
						}
					}
				},
				"tags": [
					"Authentication"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"properties": {
									"refreshToken": {
										"type": "string"
									}
								},
								"required": [
									"refreshToken"
								],
								"type": "object"
							}
						}
					}
				}
			}
		},
		"/api/auth/logout": {
			"post": {
				"operationId": "Logout",
				"responses": {
					"200": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ServiceResponse_null_"
								}
							}
						}
					}
				},
				"tags": [
					"Authentication"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": []
			}
		},
...
	},
	"servers": [
		{
			"url": "/"
		}
	]
}